[REF]l10n_ve_stock_account,l10n_ve_donation:Se mueven flujo de donaciones a nuevo modulo
Se necesita que el modulo de donaciones sea independiente del modulo de
l10n_ve_stock_account,para esto se mueven sus flujo a el nuevo modulo
llamado Venezuela Donaciones.

Link:https://www.binauraldev.com/odoo/action-341/67625
