+++ l10n_ve_donation/__init__.py
@@ -1,0 +1,1 @@
+from . import models
