+++ l10n_ve_donation/__manifest__.py
@@ -1,0 +1,25 @@
+{
+    "name": "Venezuela - Donaciones",
+    "version": "19.0.1.0.0",
+    "category": "Accounting/Accounting",
+    "summary": "Venezuela - Donaciones",
+    "author": "",
+    "website": "",
+    "depends": ["l10n_ve_accountant", "l10n_ve_stock_account"],
+    "data": [
+        "data/donation_certificate_paperformat.xml",
+        "report/donation_certificate_report.xml",
+        "report/donation_certificate_template.xml",
+        "views/account_move_views.xml",
+        "views/product_template_views.xml",
+        "views/res_config_setting_views.xml",
+        "views/sale_order_views.xml",
+        "views/stock_picking_type_views.xml",
+        "views/stock_picking_views.xml",
+        "views/stock_scrap_views.xml",
+        "views/stock_warehouse_views.xml",
+    ],
+    "installable": True,
+    "application": True,
+    "auto_install": False,
+}
