+++ l10n_ve_donation/models/__init__.py
@@ -1,0 +1,15 @@
+from . import (
+    account_move,
+    account_move_line,
+    product_template,
+    res_company,
+    res_config_settings,
+    sale_order,
+    stock_location,
+    stock_move,
+    stock_picking,
+    stock_picking_type,
+    stock_scrap,
+    stock_warehouse,
+)
+
