+++ l10n_ve_donation/models/stock_scrap.py
