+++ l10n_ve_donation/views/account_move_views.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="l10n_ve_donation_account_move_form_view" model="ir.ui.view">
+        <field name="name">l10n.ve.donation.account.move.form.view</field>
+        <field name="model">account.move</field>
+        <field name="inherit_id" ref="account.view_move_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//form//header" position="inside">
+                <button
+                    name="print_donation_certificate"
+                    string="Generar Certificado de Donación"
+                    type="object"
+                    class="btn-primary"
+                    invisible="not is_donation or (state != 'posted' or move_type != 'entry')"/>
+            </xpath>
+            <xpath expr="//field[@name='ref']" position="after">
+                <field name="is_donation" readonly="state != 'draft' or move_type != 'entry'"/>
+            </xpath>
+        </field>
+    </record>
+</odoo>
