+++ l10n_ve_donation/views/product_template_views.xml
@@ -1,7 +1,7 @@
 <odoo>
     <data>
-        <record id="l10n_ve_stock_account_product_template_form_view" model="ir.ui.view">
-            <field name="name">l10n_ve_stock_account.product.template.form.view</field>
+        <record id="l10n_ve_donation_product_template_form_view" model="ir.ui.view">
+            <field name="name">l10n_ve_donation.product.template.form.view</field>
             <field name="model">product.template</field>
             <field name="inherit_id" ref="product.product_template_form_view"/>
             <field name="arch" type="xml">
