+++ l10n_ve_donation/views/res_config_setting_views.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="l10n_ve_donation_inherit_l10n_ve_stock_account" model="ir.ui.view">
+        <field name="name">res.config.settings.view.form.inherit.l10n_ve_donation</field>
+        <field name="model">res.config.settings</field>
+        <field name="inherit_id" ref="l10n_ve_base.l10n_ve_base_view_res_config_settings"/>
+        <field name="arch" type="xml">
+            <xpath expr="//block[@name='l10n_ve_stock_block']" position="inside">
+                <div class="col-12 col-lg-6 o_setting_box" id="l10n_ve_settings_donation_account_id">
+                    <div class="o_setting_right_pane">
+                        <label for="donation_account_id"/>
+                        <div class="text-muted">
+                            Account to be used for donation flow.
+                        </div>
+                        <div class="content-group">
+                            <div class="mt16">
+                                <field name="donation_account_id"/>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </xpath>
+        </field>
+    </record>
+</odoo>
