+++ l10n_ve_donation/views/stock_picking_type_views.xml
@@ -1,6 +1,6 @@
 <odoo>
-    <record id="view_picking_type_form_inherit_l10n_ve_stock_account" model="ir.ui.view">
-        <field name="name">stock.picking.type.form.inherit.l10n.ve.stock.account</field>
+    <record id="view_picking_type_form_inherit_l10n_ve_donation" model="ir.ui.view">
+        <field name="name">stock.picking.type.form.inherit.l10n.ve.donation</field>
         <field name="model">stock.picking.type</field>
         <field name="inherit_id" ref="stock.view_picking_type_form"/>
         <field name="arch" type="xml">
