+++ l10n_ve_donation/views/stock_scrap_views.xml
