+++ l10n_ve_donation/views/stock_warehouse_views.xml
@@ -1,23 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="view_stock_warehouse_inherit_l10n_ve_stock_account" model="ir.ui.view">
-        <field name="name">stock.warehouse.form.inherit.l10n.ve.stock.account</field>
+    <record id="view_stock_warehouse_inherit_l10n_ve_donation" model="ir.ui.view">
+        <field name="name">stock.warehouse.form.inherit.l10n.ve.donation</field>
         <field name="model">stock.warehouse</field>
         <field name="inherit_id" ref="stock.view_warehouse"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='partner_id']" position="after">
-                <field name="readonly_is_consignation_warehouse" invisible="1"/>
-                <field name="is_consignation_warehouse"/>
-
+                <field name="readonly_is_donation_warehouse" invisible="1"/>
+                <field name="is_donation_warehouse"/>
             </xpath>
         </field>
     </record>
-    <record id="view_stock_warehouse_tree_inherit_l10n_ve_stock_account" model="ir.ui.view">
-        <field name="name">stock.warehouse.tree.inherit.l10n.ve.stock.account</field>
+    <record id="view_stock_warehouse_tree_inherit_l10n_ve_donation" model="ir.ui.view">
+        <field name="name">stock.warehouse.tree.inherit.l10n.ve.donation</field>
         <field name="model">stock.warehouse</field>
         <field name="inherit_id" ref="stock.view_warehouse_tree"/>
         <field name="arch" type="xml">
             <xpath expr="//list/field[@name='partner_id']" position="after">
-                <field name="is_consignation_warehouse" optional="hide"/>
+                <field name="is_donation_warehouse" optional="hide"/>
             </xpath>
         </field>
     </record>
