+++ l10n_ve_stock_account/__manifest__.py
@@ -13,6 +13,7 @@
         "l10n_ve_invoice",
         "l10n_ve_accountant",
         "l10n_ve_sale",
+        "l10n_ve_donation",
         "sale_stock",
         "web",
     ],
@@ -23,7 +24,7 @@
         "data/ir_cron.xml",
         "data/ir_sequence.xml",
         "data/transfer_reason.xml",
-        "data/donation_certificate_paperformat.xml",
+
         "views/account_move_views.xml",
         "wizard/picking_invoice_wizard.xml",
         "views/stock_picking_guide_dispatch_views.xml",
@@ -34,16 +35,14 @@
         "views/res_config_setting_views.xml",
         "views/stock_warehouse_views.xml",
         "views/stock_location_views.xml",
-        "views/product_template_views.xml",
-        "views/stock_picking_type.xml",
-        "views/stock_scrap_views.xml",
+
+
         "views/menuitem_views.xml",
         "views/alert_views.xml",
         "report/dispatch_guide.xml",
         "report/dispatch_guide_template.xml",
         "report/report_invoice_free_form.xml",
-        "report/donation_certificate_report.xml",
-        "report/donation_certificate_template.xml",
+
         "wizard/stock_picking_self_consumption_alert_views.xml",
     ],
     "application": True,
