+++ l10n_ve_stock_account/models/__init__.py
@@ -11,8 +11,5 @@ from . import (
     account_move_line,
     stock_move_line,
     stock_warehouse,
-    stock_location,
     product_template,
-    stock_picking_type,
-    stock_scrap,
 )
