+++ l10n_ve_stock_account/models/res_config_settings.py
@@ -37,4 +37,4 @@ class ResConfigSettings(models.TransientModel):
     hide_weight_field_dispatch_guide = fields.Boolean(
         related="company_id.hide_weight_field_dispatch_guide", readonly=False
     )
-
+
