+++ l10n_ve_stock_account/models/sale_order.py
@@ -21,7 +21,6 @@ class SaleOrder(models.Model):
         help="Document type for the sale order.",
     )
 
-    is_donation = fields.Boolean(string="Is Donation", default=False, tracking=True)
 
     is_consignation = fields.Boolean(
         string="Is Consignation",
@@ -30,11 +29,6 @@ class SaleOrder(models.Model):
         help="Indicates if this sale order is a consignation sale.",
     )
 
-    @api.onchange("is_donation")
-    def _onchange_is_donation(self):
-        if self.is_donation:
-            self.partner_id = self.company_id.partner_id
-            self.document = "invoice"
 
     ### COMPUTES ###
     @api.depends("warehouse_id", "document")
@@ -62,13 +56,7 @@ class SaleOrder(models.Model):
             self.document = self.partner_id.default_document
         else:
             self.document = "invoice"
-        if self.is_donation:
-            if self.partner_id != company_id.partner_id:
-                raise ValidationError(
-                    _(
-                        "The Contact/Customer cannot be changed when it is a donation."
-                    )
-                )
+
 
     @api.onchange("is_consignation")
     def _onchange_is_consignation(self):
@@ -101,16 +89,6 @@ class SaleOrder(models.Model):
             self.warehouse_id = warehouse_id
 
     ### CONSTRAINTS ###
-    @api.constrains("is_donation", "state")
-    def _check_is_donation(self):
-        for order in self:
-            if (order.state in ["sale", "done"]) and order._origin:
-                if order.is_donation != order._origin.is_donation:
-                    raise ValidationError(
-                        _(
-                            "The field 'Is Donation' cannot be modified on a confirmed or completed order."
-                        )
-                    )
 
     @api.constrains("warehouse_id", "order_line")
     def _check_consignation_warehouse(self):
@@ -139,8 +117,3 @@ class SaleOrder(models.Model):
                             )
                             % line.product_id.name
                         )
-
-    def _prepare_invoice(self):
-        invoice_vals = super()._prepare_invoice()
-        invoice_vals["is_donation"] = self.is_donation
-        return invoice_vals
