+++ l10n_ve_stock_account/models/stock_move.py
@@ -126,46 +126,3 @@ class StockMove(models.Model):
 
         return ves_price_unit
     
-    def _create_account_move(self):
-        """ Create account move for specific location or analytic.
-            This function is a override of the original function to add the donation logic.
-        """
-        aml_vals_list = []
-        move_to_link = set()
-        company_partner = self.env.company.partner_id
-        
-        is_donation = any(move.scrap_id and move.scrap_id.is_donation for move in self)
-        
-        for move in self:
-            if move._should_create_account_move():
-                aml_vals = move._get_account_move_line_vals()
-                if is_donation:
-                    for val in aml_vals:
-                        val["partner_id"] = company_partner.id
-                aml_vals_list += aml_vals
-                move_to_link.add(move.id)
-                
-        if not aml_vals_list:
-            return self.env['account.move']
-            
-        move_vals = {
-            "journal_id": self.company_id.account_stock_journal_id.id,
-            "line_ids": [Command.create(aml_vals) for aml_vals in aml_vals_list],
-            "date": self.env.context.get("force_period_date") or fields.Date.context_today(self),
-        }
-        
-        if is_donation:
-            reasons = [', '.join(m.scrap_id.scrap_reason_tag_ids.mapped('name')) 
-                       for m in self if m.scrap_id and m.scrap_id.is_donation and m.scrap_id.scrap_reason_tag_ids]
-            ref = ' - '.join(filter(None, reasons))
-            
-            move_vals.update({
-                "is_donation": True,
-                "partner_id": company_partner.id,
-                "ref": ref
-            })
-            
-        account_move = self.env["account.move"].sudo().create(move_vals)
-        self.env["stock.move"].browse(move_to_link).account_move_id = account_move.id
-        account_move._post()
-        return account_move
