+++ l10n_ve_stock_account/views/res_config_setting_views.xml
@@ -108,19 +108,6 @@
                         </div>
                     </div>
                 </div>
-                <div class="col-12 col-lg-6 o_setting_box" id="l10n_ve_settings_donation_account_id">
-                    <div class="o_setting_right_pane">
-                        <label for="donation_account_id"/>
-                        <div class="text-muted">
-                            Account to be used for donation flow.
-                        </div>
-                        <div class="content-group">
-                            <div class="mt16">
-                                <field name="donation_account_id"/>
-                            </div>
-                        </div>
-                    </div>
-                </div>
             </xpath>
         </field>
     </record>
